+++ .atomist/editors/HelloTypeScript.ts
@@ -1,15 +1,25 @@
-import { ProjectEditor } from '@atomist/rug/operations/ProjectEditor'
+import { EditProject } from '@atomist/rug/operations/ProjectEditor'
 import { Project } from '@atomist/rug/model/Core'
-import { Result, Status } from '@atomist/rug/operations/RugOperation'
+import { Pattern } from '@atomist/rug/operations/RugOperation'
+import { Editor, Parameter, Tags } from '@atomist/rug/operations/Decorators'
 
-class HelloTypeScript implements ProjectEditor {
-    tags: string[] = ["documentation"]
-    name: string = "HelloTypeScript"
-    description: string = "Added to a project by AddTypeScriptEditor"
-    edit(project: Project): Result {
-       project.addFile("hello.txt", "Hello, World!")
-       return new Result(Status.Success, "I made a file")
+@Editor("HelloTypeScript", "@DESCRIPTION@")
+@Tags("documentation")
+class HelloTypeScript implements EditProject {
+
+    @Parameter({
+        displayName: "Some Input",
+        description: "Example of how to specify a parameter using decorators",
+        pattern: Pattern.any,
+        validInput: "A description of the valid input",
+        minLength: 1,
+        maxLength: 100
+    })
+    input_parameter: string;
+
+    edit(project: Project) {
+        project.addFile("hello.txt", "Hello, World!\n" + this.input_parameter + "\n");
     }
 }
 
-export let hello = new HelloTypeScript()
+export let helloTypeScript = new HelloTypeScript()